Move conftest out to tests/gui
