+++ tests/gui/models/test_progress_proxy.py
@@ -6,7 +6,7 @@ from ert_shared.status.entity.state import (
 from ert_shared.ensemble_evaluator.entity.snapshot import PartialSnapshot, Realization
 from PyQt5.QtCore import QModelIndex
 from ert_gui.model.progress_proxy import ProgressProxyModel
-from tests.gui.models.conftest import partial_snapshot
+from tests.gui.conftest import partial_snapshot
 from ert_gui.model.snapshot import ProgressRole, SnapshotModel
 from pytestqt.qt_compat import qt_api
