+++ tests/gui/models/test_snapshot_model.py
@@ -1,7 +1,14 @@
-from PyQt5.QtCore import QModelIndex
-from tests.gui.models.conftest import partial_snapshot
-from ert_gui.model.snapshot import NodeRole, SnapshotModel
+from ert_gui.model.snapshot import RealJobColorHint, SnapshotModel
+from ert_shared.ensemble_evaluator.entity.snapshot import Job, PartialSnapshot
+from ert_shared.status.entity.state import (
+    COLOR_PENDING,
+    COLOR_RUNNING,
+    JOB_STATE_RUNNING,
+)
 from pytestqt.qt_compat import qt_api
+from qtpy.QtCore import QModelIndex
+from qtpy.QtGui import QColor
+from tests.gui.conftest import partial_snapshot
 
 
 def test_using_qt_model_tester(qtmodeltester, full_snapshot):
